Improve look of the user email settings page
This commit improves the look of the users email settings page.
It highlights the current menu item, aligns the formatting
of the list with the other pages and introduces sorting
for the email adresses.

The primary address should always be shown first with the others
coming after in lexical sorted order.